backend/requirements.txt
asarPy==1.0.1
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.0
Django==5.2.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
